dwarf: handle variable definitions that appear before specification
If a DW_TAG_variable has a DW_AT_specification that refers to a later
DIE, then we failed to inherit the attributes from the specification.

Fix by deferring this inheritance until after all variables are parsed.

Note that methods may have a similar problem, but they are harder to fix
due to their children needing to inherit the namespace too, so haven't
done that yet.