Fix padding for structs with members of size 0
These occur often in rust. The problem is that when calculating
padding we weren't sorting the size 0 member before the member
at the same offset. The catch was that to fix the sorting we
needed to finish parsing the unit so that we can calculate type
sizes. So we now calculate padding when printing, instead of when
parsing.